+++ lib/c/states.dart
@@ -1353,8 +1353,7 @@ class Running extends ControllerState {
     }
   }
 
-  Future<void> runProgramLoop(
-      {Set<int> acceptableErrors = const <int>{}}) async {
+  Future<void> runProgramLoop() async {
     final ProgramListener listener = model.memory.program.programListener;
     final settings = controller.model.settings;
     final program = model.memory.program;
@@ -1472,10 +1471,6 @@ class Running extends ControllerState {
         break;
       } else if (err != null) {
         assert(_pushedRunner == null);
-        if (acceptableErrors.contains(err.num15)) {
-          _fake.pendingError = null;
-          throw err;
-        }
         listener.onError(err);
         _onDone();
         _stopNext = false;